+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 This file is used to list changes made in each version of cookbook-openstack-identity.
 ## 10.0.1
 * Update keystone.conf from mode 0644 to 0640
+* Allow hash_algorithm to be configurable
 
 ## 10.0.0
 * Upgrading to Juno